Move book file entity
