+++ src-tauri/libcalibre/src/client.rs
@@ -1,9 +1,9 @@
 use crate::cover_image::cover_image_data_from_path;
-use crate::domain::book_file::entity::NewBookFile;
 use crate::dtos::file::NewFileDto;
 use crate::dtos::library::NewLibraryEntryDto;
 use crate::dtos::library::NewLibraryFileDto;
 use crate::dtos::library::UpdateLibraryEntryDto;
+use crate::entities::book_file::NewBookFile;
 use crate::Book;
 use chrono::DateTime;
 use chrono::NaiveDateTime;
