--- src-tauri/libcalibre/src/domain/book_file/mod.rs
@@ -1,1 +1,0 @@
-pub mod entity;
