--- src-tauri/libcalibre/src/domain/mod.rs
@@ -1,1 +1,0 @@
-pub mod book_file;
