+++ src-tauri/libcalibre/src/entities.rs
@@ -1,3 +1,4 @@
 pub mod author;
 pub mod book;
 pub mod book_aggregate;
+pub mod book_file;
